fix(test): cast carton_count comparison to bigint for pgTAP is()
carton_count is bigint (count(*) in the view); pgTAP's is() has no
(bigint, integer, unknown) overload, so the literal must be cast.

+++ supabase/tests/dispatch_shipment_scoped_authority.test.sql
@@ -119,7 +119,7 @@ select is(
 );
 select is(
   (select carton_count from public.b2b_dispatch_shipment_execution_view where consignment_id = '99600000-0000-0000-0000-000000000001'),
-  1, 'carton count is available (1)'
+  1::bigint, 'carton count is available (1)'
 );
 select is(
   (select transporter_name from public.b2b_dispatch_shipment_execution_view where consignment_id = '99600000-0000-0000-0000-000000000001'),
